Add .long() to targets

+++ model.py
@@ -53,7 +53,7 @@ class MultiHotCollator():
         for i, t in enumerate(batch):
             x[i].scatter_(0, t[0], 1.0)
         
-        y = torch.stack([t[1] for t in batch])
+        y = torch.stack([t[1] for t in batch]).long()
 
         return x, y
